+++ dataworkspace/dataworkspace/apps/explorer/utils.py
@@ -1,13 +1,24 @@
 import re
+from contextlib import contextmanager
+from datetime import timedelta
 
-
+import psycopg2
 import sqlparse
+from six import text_type
+
 from django.conf import settings
 from django.contrib.auth import REDIRECT_FIELD_NAME
 from django.contrib.auth.forms import AuthenticationForm
 from django.contrib.auth.views import LoginView
-from six import text_type
+from django.core.cache import cache
 
+from dataworkspace.apps.core.models import Database
+from dataworkspace.apps.core.utils import (
+    new_private_database_credentials,
+    source_tables_for_user,
+    db_role_schema_suffix_for_user,
+    postgres_user,
+)
 from dataworkspace.apps.explorer import app_settings
 
 EXPLORER_PARAM_TOKEN = "$$"
@@ -125,20 +136,65 @@ class InvalidExplorerConnectionException(Exception):
     pass
 
 
-def get_valid_connection(alias=None):
+def get_user_explorer_connection_settings(user, alias):
     from dataworkspace.apps.explorer.connections import (  # pylint: disable=import-outside-toplevel
         connections,
     )
 
     if not alias:
-        return connections[settings.EXPLORER_DEFAULT_CONNECTION]
+        alias = settings.EXPLORER_DEFAULT_CONNECTION
 
     if alias not in connections:
         raise InvalidExplorerConnectionException(
             'Attempted to access connection %s, but that is not a registered Explorer connection.'
             % alias
         )
-    return connections[alias]
+
+    def get_available_user_connections(_user_credentials):
+        return {data['memorable_name']: data for data in _user_credentials}
+
+    cache_key = f"explorer_credentials_{user.profile.sso_id}"
+    user_credentials = cache.get(cache_key, None)
+
+    if not user_credentials:
+        db_role_schema_suffix = db_role_schema_suffix_for_user(user)
+        source_tables = source_tables_for_user(user)
+        db_user = postgres_user(user.email, suffix='explorer')
+        duration = timedelta(hours=24)
+        cache_duration = (duration - timedelta(minutes=15)).total_seconds()
+
+        user_credentials = new_private_database_credentials(
+            db_role_schema_suffix,
+            source_tables,
+            db_user,
+            valid_for=duration,
+            force_create_for_databases=Database.objects.filter(
+                memorable_name__in=connections.keys()
+            ).all(),
+        )
+        cache.set(cache_key, user_credentials, timeout=cache_duration)
+
+    db_aliases_to_credentials = get_available_user_connections(user_credentials)
+    if alias not in db_aliases_to_credentials:
+        raise RuntimeError(
+            f"The credentials for {user.email} did not include any for the `{alias}` database."
+        )
+
+    return db_aliases_to_credentials[alias]
+
+
+@contextmanager
+def user_explorer_connection(user, alias=None):
+    connection_settings = get_user_explorer_connection_settings(user, alias)
+
+    with psycopg2.connect(
+        dbname=connection_settings['db_name'],
+        host=connection_settings['db_host'],
+        user=connection_settings['db_user'],
+        password=connection_settings['db_password'],
+        port=connection_settings['db_port'],
+    ) as conn:
+        yield conn
 
 
 def get_total_pages(total_rows, page_size):