+++ dataworkspace/dataworkspace/tests/explorer/test_tasks.py
@@ -1,14 +1,9 @@
 from datetime import datetime, timedelta
-from unittest.mock import patch
 
 from django.test import TestCase
 
-from dataworkspace.apps.explorer.app_settings import EXPLORER_DEFAULT_CONNECTION as CONN
 from dataworkspace.apps.explorer.models import QueryLog
-from dataworkspace.apps.explorer.tasks import (
-    build_schema_cache_async,
-    truncate_querylogs,
-)
+from dataworkspace.apps.explorer.tasks import truncate_querylogs
 
 
 class TestTasks(TestCase):
@@ -23,11 +18,3 @@ class TestTasks(TestCase):
         )
         truncate_querylogs(30)
         self.assertEqual(QueryLog.objects.count(), 1)
-
-    @patch('dataworkspace.apps.explorer.schema.build_schema_info')
-    @patch('dataworkspace.apps.explorer.schema.cache.set')
-    def test_build_schema_cache_async(self, _, mocked_build):
-        mocked_build.return_value = ['list_of_tuples']
-        schema = build_schema_cache_async(CONN)
-        assert mocked_build.called
-        self.assertEqual(schema, ['list_of_tuples'])